Fixed ACT-1405 to save CustomServiceTask ID in model file.
CustomServiceTasks are now once more matched to their extension CST by 
inspecting a property that is explictly marshalled in the model's 
BPMN document. The attribute (XML: activiti:extensionId) is read when
parsing the model and saved when exporting it. It is also created for
newly added nodes, if they are added to the canvas from an extension.

This fix has consequences for existing models. The @Runtime annotation
is no longer matched to the activiti:class attribute in the BPMN
document, because it cannot be relied on. This means that opening a
diagram with CSTs will no longer recognize the nodes as
CustomServiceTasks. The BPMN must be manually adjusted. To do this,
users should add the activiti:extensionId attribute to existing service
tasks. The value should be the fully qualified name of the CST extension
class. For example:
activiti:extensionId="org.acme.servicetasks.AcmeMoneyTask"