* Added NATURE_ID to ActivitiConstants.java in utility project and corrected various references to use it.
* Finished implementation of PropertyCallActivitySection.java:
** Automatically checks the process ID and if found in any open diagram
in any open project enables the appropriate button to open the diagram
** In case the user selects the button to open the diagram but the ID is
ambiguous, a dialog opens to select the diagram to open
** Added another button to select a process by process ID from any
available open project

* Added context menu feature OpenCalledElementForCallActivity.java and
introduced it in the ActivitiToolBehaviourProvider.java. It allows to
open the diagram referenced by the process ID given in the Call
Activities 'Called Element' property in case it is unique (the
non-unique dialog is currently only available via property sheet).