Completed for ACT-1572.
* Refactored code provided by Michael Priess to support all three output
options and changed the @Runtime attribute names to more closely reflect
the terminology Activiti uses in the output and user guide. This will
break any existing extensions when people upgrade, but changes are
limited to a new and better name.
* Made sure the properties section is not shown if the implementation
type
is expression. The section as a whole will still be shown, because there
may be @Help annotations needed to be shown, but @Property annotations
will not be processed into fields.
* Removed the duplicate setting of the implementation type and
implementation string. This is already present in the model when loading
and handles all options (java delegate, expression, java delegate
expression) gracefully.
* Added examples of CustomServiceTask for the new @Runtime
configurations. Also added a project containing some example delegate
classes.